+++ cli/compat/esm_resolver.rs
@@ -4,9 +4,6 @@ use super::errors;
 use crate::resolver::ImportMapResolver;
 use deno_core::error::generic_error;
 use deno_core::error::AnyError;
-use deno_core::serde_json;
-use deno_core::serde_json::Map;
-use deno_core::serde_json::Value;
 use deno_core::url::Url;
 use deno_core::ModuleSpecifier;
 use deno_graph::source::ResolveResponse;
@@ -14,6 +11,9 @@ use deno_graph::source::Resolver;
 use regex::Regex;
 use std::path::PathBuf;
 
+use super::package_json::get_package_config;
+use super::package_json::get_package_scope_config;
+use super::package_json::PackageConfig;
 #[derive(Debug, Default)]
 pub(crate) struct NodeEsmResolver {
   maybe_import_map_resolver: Option<ImportMapResolver>,
@@ -207,7 +207,9 @@ fn module_resolve(
   let resolved = if should_be_treated_as_relative_or_absolute_path(specifier) {
     base.join(specifier)?
   } else if specifier.starts_with('#') {
-    package_imports_resolve(specifier, base, conditions)?
+    super::conditional_exports::package_imports_resolve(
+      specifier, base, conditions,
+    )?
   } else if let Ok(resolved) = Url::parse(specifier) {
     resolved
   } else {
@@ -265,532 +267,7 @@ fn finalize_resolution(
   Ok(resolved)
 }
 
-fn throw_import_not_defined(
-  specifier: &str,
-  package_json_url: Option<ModuleSpecifier>,
-  base: &ModuleSpecifier,
-) -> AnyError {
-  errors::err_package_import_not_defined(
-    specifier,
-    package_json_url.map(|u| to_file_path_string(&u.join(".").unwrap())),
-    &to_file_path_string(base),
-  )
-}
-
-fn pattern_key_compare(a: &str, b: &str) -> i32 {
-  let a_pattern_index = a.find('*');
-  let b_pattern_index = b.find('*');
-
-  let base_len_a = if let Some(index) = a_pattern_index {
-    index + 1
-  } else {
-    a.len()
-  };
-  let base_len_b = if let Some(index) = b_pattern_index {
-    index + 1
-  } else {
-    b.len()
-  };
-
-  if base_len_a > base_len_b {
-    return -1;
-  }
-
-  if base_len_b > base_len_a {
-    return 1;
-  }
-
-  if a_pattern_index.is_none() {
-    return 1;
-  }
-
-  if b_pattern_index.is_none() {
-    return -1;
-  }
-
-  if a.len() > b.len() {
-    return -1;
-  }
-
-  if b.len() > a.len() {
-    return 1;
-  }
-
-  0
-}
-
-fn package_imports_resolve(
-  name: &str,
-  base: &ModuleSpecifier,
-  conditions: &[&str],
-) -> Result<ModuleSpecifier, AnyError> {
-  if name == "#" || name.starts_with("#/") || name.ends_with('/') {
-    let reason = "is not a valid internal imports specifier name";
-    return Err(errors::err_invalid_module_specifier(
-      name,
-      reason,
-      Some(to_file_path_string(base)),
-    ));
-  }
-
-  let mut package_json_url = None;
-
-  let package_config = get_package_scope_config(base)?;
-  if package_config.exists {
-    package_json_url =
-      Some(Url::from_file_path(package_config.pjsonpath).unwrap());
-    if let Some(imports) = &package_config.imports {
-      if imports.contains_key(name) && !name.contains('*') {
-        let maybe_resolved = resolve_package_target(
-          package_json_url.clone().unwrap(),
-          imports.get(name).unwrap().to_owned(),
-          "".to_string(),
-          name.to_string(),
-          base,
-          false,
-          true,
-          conditions,
-        )?;
-        if let Some(resolved) = maybe_resolved {
-          return Ok(resolved);
-        }
-      } else {
-        let mut best_match = "";
-        let mut best_match_subpath = None;
-        for key in imports.keys() {
-          let pattern_index = key.find('*');
-          if let Some(pattern_index) = pattern_index {
-            let key_sub = &key[0..=pattern_index];
-            if name.starts_with(key_sub) {
-              let pattern_trailer = &key[pattern_index + 1..];
-              if name.len() > key.len()
-                && name.ends_with(&pattern_trailer)
-                && pattern_key_compare(best_match, key) == 1
-                && key.rfind('*') == Some(pattern_index)
-              {
-                best_match = key;
-                best_match_subpath = Some(
-                  name[pattern_index..=(name.len() - pattern_trailer.len())]
-                    .to_string(),
-                );
-              }
-            }
-          }
-        }
-
-        if !best_match.is_empty() {
-          let target = imports.get(best_match).unwrap().to_owned();
-          let maybe_resolved = resolve_package_target(
-            package_json_url.clone().unwrap(),
-            target,
-            best_match_subpath.unwrap(),
-            best_match.to_string(),
-            base,
-            true,
-            true,
-            conditions,
-          )?;
-          if let Some(resolved) = maybe_resolved {
-            return Ok(resolved);
-          }
-        }
-      }
-    }
-  }
-
-  Err(throw_import_not_defined(name, package_json_url, base))
-}
-
-fn is_conditional_exports_main_sugar(
-  exports: &Value,
-  package_json_url: &ModuleSpecifier,
-  base: &ModuleSpecifier,
-) -> Result<bool, AnyError> {
-  if exports.is_string() || exports.is_array() {
-    return Ok(true);
-  }
-
-  if exports.is_null() || !exports.is_object() {
-    return Ok(false);
-  }
-
-  let exports_obj = exports.as_object().unwrap();
-  let mut is_conditional_sugar = false;
-  let mut i = 0;
-  for key in exports_obj.keys() {
-    let cur_is_conditional_sugar = key.is_empty() || !key.starts_with('.');
-    if i == 0 {
-      is_conditional_sugar = cur_is_conditional_sugar;
-      i += 1;
-    } else if is_conditional_sugar != cur_is_conditional_sugar {
-      return Err(errors::err_invalid_package_config(
-        &to_file_path_string(package_json_url),
-        Some(base.as_str().to_string()),
-        Some("\"exports\" cannot contains some keys starting with \'.\' and some not.
-        The exports object must either be an object of package subpath keys
-        or an object of main entry condition name keys only.".to_string())
-      ));
-    }
-  }
-
-  Ok(is_conditional_sugar)
-}
-
-fn throw_invalid_package_target(
-  subpath: String,
-  target: String,
-  package_json_url: &ModuleSpecifier,
-  internal: bool,
-  base: &ModuleSpecifier,
-) -> AnyError {
-  errors::err_invalid_package_target(
-    to_file_path_string(&package_json_url.join(".").unwrap()),
-    subpath,
-    target,
-    internal,
-    Some(base.as_str().to_string()),
-  )
-}
-
-fn throw_invalid_subpath(
-  subpath: String,
-  package_json_url: &ModuleSpecifier,
-  internal: bool,
-  base: &ModuleSpecifier,
-) -> AnyError {
-  let ie = if internal { "imports" } else { "exports" };
-  let reason = format!(
-    "request is not a valid subpath for the \"{}\" resolution of {}",
-    ie,
-    to_file_path_string(package_json_url)
-  );
-  errors::err_invalid_module_specifier(
-    &subpath,
-    &reason,
-    Some(to_file_path_string(base)),
-  )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn resolve_package_target_string(
-  target: String,
-  subpath: String,
-  match_: String,
-  package_json_url: ModuleSpecifier,
-  base: &ModuleSpecifier,
-  pattern: bool,
-  internal: bool,
-  conditions: &[&str],
-) -> Result<ModuleSpecifier, AnyError> {
-  if !subpath.is_empty() && !pattern && !target.ends_with('/') {
-    return Err(throw_invalid_package_target(
-      match_,
-      target,
-      &package_json_url,
-      internal,
-      base,
-    ));
-  }
-
-  let invalid_segment_re =
-    Regex::new(r"(^|\|/)(..?|node_modules)(\|/|$)").expect("bad regex");
-  let pattern_re = Regex::new(r"\*").expect("bad regex");
-
-  if !target.starts_with("./") {
-    if internal && !target.starts_with("../") && !target.starts_with('/') {
-      let is_url = Url::parse(&target).is_ok();
-      if !is_url {
-        let export_target = if pattern {
-          pattern_re
-            .replace(&target, |_caps: &regex::Captures| subpath.clone())
-            .to_string()
-        } else {
-          format!("{}{}", target, subpath)
-        };
-        return package_resolve(&export_target, &package_json_url, conditions);
-      }
-    }
-    return Err(throw_invalid_package_target(
-      match_,
-      target,
-      &package_json_url,
-      internal,
-      base,
-    ));
-  }
-
-  if invalid_segment_re.is_match(&target[2..]) {
-    return Err(throw_invalid_package_target(
-      match_,
-      target,
-      &package_json_url,
-      internal,
-      base,
-    ));
-  }
-
-  let resolved = package_json_url.join(&target)?;
-  let resolved_path = resolved.path();
-  let package_url = package_json_url.join(".").unwrap();
-  let package_path = package_url.path();
-
-  if !resolved_path.starts_with(package_path) {
-    return Err(throw_invalid_package_target(
-      match_,
-      target,
-      &package_json_url,
-      internal,
-      base,
-    ));
-  }
-
-  if subpath.is_empty() {
-    return Ok(resolved);
-  }
-
-  if invalid_segment_re.is_match(&subpath) {
-    let request = if pattern {
-      match_.replace("*", &subpath)
-    } else {
-      format!("{}{}", match_, subpath)
-    };
-    return Err(throw_invalid_subpath(
-      request,
-      &package_json_url,
-      internal,
-      base,
-    ));
-  }
-
-  if pattern {
-    let replaced = pattern_re
-      .replace(resolved.as_str(), |_caps: &regex::Captures| subpath.clone());
-    let url = Url::parse(&replaced)?;
-    return Ok(url);
-  }
-
-  Ok(resolved.join(&subpath)?)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn resolve_package_target(
-  package_json_url: ModuleSpecifier,
-  target: Value,
-  subpath: String,
-  package_subpath: String,
-  base: &ModuleSpecifier,
-  pattern: bool,
-  internal: bool,
-  conditions: &[&str],
-) -> Result<Option<ModuleSpecifier>, AnyError> {
-  if let Some(target) = target.as_str() {
-    return Ok(Some(resolve_package_target_string(
-      target.to_string(),
-      subpath,
-      package_subpath,
-      package_json_url,
-      base,
-      pattern,
-      internal,
-      conditions,
-    )?));
-  } else if let Some(target_arr) = target.as_array() {
-    if target_arr.is_empty() {
-      return Ok(None);
-    }
-
-    let mut last_error = None;
-    for target_item in target_arr {
-      let resolved_result = resolve_package_target(
-        package_json_url.clone(),
-        target_item.to_owned(),
-        subpath.clone(),
-        package_subpath.clone(),
-        base,
-        pattern,
-        internal,
-        conditions,
-      );
-
-      if let Err(e) = resolved_result {
-        let err_string = e.to_string();
-        last_error = Some(e);
-        if err_string.starts_with("[ERR_INVALID_PACKAGE_TARGET]") {
-          continue;
-        }
-        return Err(last_error.unwrap());
-      }
-      let resolved = resolved_result.unwrap();
-      if resolved.is_none() {
-        last_error = None;
-        continue;
-      }
-      return Ok(resolved);
-    }
-    if last_error.is_none() {
-      return Ok(None);
-    }
-    return Err(last_error.unwrap());
-  } else if let Some(target_obj) = target.as_object() {
-    for key in target_obj.keys() {
-      // TODO(bartlomieju): verify that keys are not numeric
-      // return Err(errors::err_invalid_package_config(
-      //   to_file_path_string(package_json_url),
-      //   Some(base.as_str().to_string()),
-      //   Some("\"exports\" cannot contain numeric property keys.".to_string()),
-      // ));
-
-      if key == "default" || conditions.contains(&key.as_str()) {
-        let condition_target = target_obj.get(key).unwrap().to_owned();
-        let resolved = resolve_package_target(
-          package_json_url.clone(),
-          condition_target,
-          subpath.clone(),
-          package_subpath.clone(),
-          base,
-          pattern,
-          internal,
-          conditions,
-        )?;
-        if resolved.is_none() {
-          continue;
-        }
-        return Ok(resolved);
-      }
-    }
-  } else if target.is_null() {
-    return Ok(None);
-  }
-
-  Err(throw_invalid_package_target(
-    package_subpath,
-    target.to_string(),
-    &package_json_url,
-    internal,
-    base,
-  ))
-}
-
-fn throw_exports_not_found(
-  subpath: String,
-  package_json_url: &ModuleSpecifier,
-  base: &ModuleSpecifier,
-) -> AnyError {
-  errors::err_package_path_not_exported(
-    to_file_path_string(&package_json_url.join(".").unwrap()),
-    subpath,
-    Some(to_file_path_string(base)),
-  )
-}
-
-fn package_exports_resolve(
-  package_json_url: ModuleSpecifier,
-  package_subpath: String,
-  package_config: PackageConfig,
-  base: &ModuleSpecifier,
-  conditions: &[&str],
-) -> Result<ModuleSpecifier, AnyError> {
-  let exports = &package_config.exports.unwrap();
-
-  let exports_map =
-    if is_conditional_exports_main_sugar(exports, &package_json_url, base)? {
-      let mut map = Map::new();
-      map.insert(".".to_string(), exports.to_owned());
-      map
-    } else {
-      exports.as_object().unwrap().to_owned()
-    };
-
-  if exports_map.contains_key(&package_subpath)
-    && package_subpath.find('*').is_none()
-    && !package_subpath.ends_with('/')
-  {
-    let target = exports_map.get(&package_subpath).unwrap().to_owned();
-    let resolved = resolve_package_target(
-      package_json_url.clone(),
-      target,
-      "".to_string(),
-      package_subpath.to_string(),
-      base,
-      false,
-      false,
-      conditions,
-    )?;
-    if resolved.is_none() {
-      return Err(throw_exports_not_found(
-        package_subpath,
-        &package_json_url,
-        base,
-      ));
-    }
-    return Ok(resolved.unwrap());
-  }
-
-  let mut best_match = "";
-  let mut best_match_subpath = None;
-  for key in exports_map.keys() {
-    let pattern_index = key.find('*');
-    if let Some(pattern_index) = pattern_index {
-      let key_sub = &key[0..=pattern_index];
-      if package_subpath.starts_with(key_sub) {
-        // When this reaches EOL, this can throw at the top of the whole function:
-        //
-        // if (StringPrototypeEndsWith(packageSubpath, '/'))
-        //   throwInvalidSubpath(packageSubpath)
-        //
-        // To match "imports" and the spec.
-        if package_subpath.ends_with('/') {
-          // TODO(bartlomieju):
-          // emitTrailingSlashPatternDeprecation();
-        }
-        let pattern_trailer = &key[pattern_index + 1..];
-        if package_subpath.len() > key.len()
-          && package_subpath.ends_with(&pattern_trailer)
-          && pattern_key_compare(best_match, key) == 1
-          && key.rfind('*') == Some(pattern_index)
-        {
-          best_match = key;
-          best_match_subpath = Some(
-            package_subpath
-              [pattern_index..=(package_subpath.len() - pattern_trailer.len())]
-              .to_string(),
-          );
-        }
-      }
-    }
-  }
-
-  if !best_match.is_empty() {
-    let target = exports.get(best_match).unwrap().to_owned();
-    let maybe_resolved = resolve_package_target(
-      package_json_url.clone(),
-      target,
-      best_match_subpath.unwrap(),
-      best_match.to_string(),
-      base,
-      true,
-      false,
-      conditions,
-    )?;
-    if let Some(resolved) = maybe_resolved {
-      return Ok(resolved);
-    } else {
-      return Err(throw_exports_not_found(
-        package_subpath,
-        &package_json_url,
-        base,
-      ));
-    }
-  }
-
-  Err(throw_exports_not_found(
-    package_subpath,
-    &package_json_url,
-    base,
-  ))
-}
-
-fn package_resolve(
+pub(crate) fn package_resolve(
   specifier: &str,
   base: &ModuleSpecifier,
   conditions: &[&str],
@@ -806,7 +283,7 @@ fn package_resolve(
     if package_config.name.as_ref() == Some(&package_name) {
       if let Some(exports) = &package_config.exports {
         if !exports.is_null() {
-          return package_exports_resolve(
+          return super::conditional_exports::package_exports_resolve(
             package_json_url,
             package_subpath,
             package_config,
@@ -855,7 +332,7 @@ fn package_resolve(
     let package_config =
       get_package_config(package_json_path.clone(), specifier, Some(base))?;
     if package_config.exports.is_some() {
-      return package_exports_resolve(
+      return super::conditional_exports::package_exports_resolve(
         package_json_url,
         package_subpath,
         package_config,
@@ -934,17 +411,6 @@ fn parse_package_name(
   Ok((package_name, package_subpath, is_scoped))
 }
 
-#[derive(Clone, Debug)]
-struct PackageConfig {
-  exists: bool,
-  exports: Option<Value>,
-  imports: Option<Map<String, Value>>,
-  main: Option<String>,
-  name: Option<String>,
-  pjsonpath: PathBuf,
-  typ: String,
-}
-
 pub fn check_if_should_use_esm_loader(
   main_module: &ModuleSpecifier,
 ) -> Result<bool, AnyError> {
@@ -957,146 +423,6 @@ pub fn check_if_should_use_esm_loader(
   }
   let package_config = get_package_scope_config(main_module)?;
   Ok(package_config.typ == "module")
-}
-
-fn get_package_config(
-  path: PathBuf,
-  specifier: &str,
-  maybe_base: Option<&ModuleSpecifier>,
-) -> Result<PackageConfig, AnyError> {
-  // TODO(bartlomieju):
-  // if let Some(existing) = package_json_cache.get(path) {
-  //   return Ok(existing.clone());
-  // }
-
-  let result = std::fs::read_to_string(&path);
-
-  let source = result.unwrap_or_else(|_| "".to_string());
-  if source.is_empty() {
-    let package_config = PackageConfig {
-      pjsonpath: path,
-      exists: false,
-      main: None,
-      name: None,
-      typ: "none".to_string(),
-      exports: None,
-      imports: None,
-    };
-    // TODO(bartlomieju):
-    // package_json_cache.set(package_json_path, package_config.clone());
-    return Ok(package_config);
-  }
-
-  let package_json: Value = serde_json::from_str(&source).map_err(|err| {
-    let base_msg = maybe_base.map(|base| {
-      format!("\"{}\" from {}", specifier, to_file_path(base).display())
-    });
-    errors::err_invalid_package_config(
-      &path.display().to_string(),
-      base_msg,
-      Some(err.to_string()),
-    )
-  })?;
-
-  let imports_val = package_json.get("imports");
-  let main_val = package_json.get("main");
-  let name_val = package_json.get("name");
-  let typ_val = package_json.get("type");
-  let exports = package_json.get("exports").map(|e| e.to_owned());
-
-  let imports = if let Some(imp) = imports_val {
-    imp.as_object().map(|imp| imp.to_owned())
-  } else {
-    None
-  };
-  let main = if let Some(m) = main_val {
-    m.as_str().map(|m| m.to_string())
-  } else {
-    None
-  };
-  let name = if let Some(n) = name_val {
-    n.as_str().map(|n| n.to_string())
-  } else {
-    None
-  };
-
-  // Ignore unknown types for forwards compatibility
-  let typ = if let Some(t) = typ_val {
-    if let Some(t) = t.as_str() {
-      if t != "module" && t != "commonjs" {
-        "none".to_string()
-      } else {
-        t.to_string()
-      }
-    } else {
-      "none".to_string()
-    }
-  } else {
-    "none".to_string()
-  };
-
-  let package_config = PackageConfig {
-    pjsonpath: path,
-    exists: true,
-    main,
-    name,
-    typ,
-    exports,
-    imports,
-  };
-  // TODO(bartlomieju):
-  // package_json_cache.set(package_json_path, package_config.clone());
-  Ok(package_config)
-}
-
-fn get_package_scope_config(
-  resolved: &ModuleSpecifier,
-) -> Result<PackageConfig, AnyError> {
-  let mut package_json_url = resolved.join("./package.json")?;
-
-  loop {
-    let package_json_path = package_json_url.path();
-
-    if package_json_path.ends_with("node_modules/package.json") {
-      break;
-    }
-
-    let package_config = get_package_config(
-      to_file_path(&package_json_url),
-      resolved.as_str(),
-      None,
-    )?;
-
-    if package_config.exists {
-      return Ok(package_config);
-    }
-
-    let last_package_json_url = package_json_url.clone();
-    package_json_url = package_json_url.join("../package.json")?;
-
-    // TODO(bartlomieju): I'm not sure this will work properly
-    // Terminates at root where ../package.json equals ../../package.json
-    // (can't just check "/package.json" for Windows support)
-    if package_json_url.path() == last_package_json_url.path() {
-      break;
-    }
-  }
-
-  let package_json_path = to_file_path(&package_json_url);
-  let package_config = PackageConfig {
-    pjsonpath: package_json_path,
-    exists: false,
-    main: None,
-    name: None,
-    typ: "none".to_string(),
-    exports: None,
-    imports: None,
-  };
-
-  // TODO(bartlomieju):
-  // package_json_cache.set(package_json_path, package_config.clone());
-
-  Ok(package_config)
 }
 
 fn file_exists(path_url: &ModuleSpecifier) -> bool {
