+++ cli/compat/mod.rs
@@ -1,7 +1,9 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
+mod conditional_exports;
 mod errors;
 mod esm_resolver;
+mod package_json;
 
 use deno_core::error::AnyError;
 use deno_core::located_script_name;
